On displays > 1024px, show navbar at bottom of screen
Modified header height to subtract the width of navbar. Adjusted line-height of navbar to directly center links horizontally. Added target attr to links that open to external site so that a new tab will be opened instead of navigating user from the main site. Removed 'Forum' link from 'More' menu so that the user does not have to drill down into popover. Add JS/CSS to handle when navbar is too close to bottom, popover will open above the link, otherwise popover will continue to open below the link. Updated arrow CSS so that it does not overlap with navbar.